Move navbar from header to separate template, generate once and serve with SSI
We are making the assumption the generated template will be
available at URI /navbar.html. It would be better to use R_PATH template
variable and include file (vs. virtual) but at least nginx rejects template
variable relative URIs pointing to a parent path as unsafe.

It would also be better if a config parameter, possibly linked to
html_strategy, was exposed as a template variable, in order to drive a
decision whether to render the navbar in head.tmpl or to use SSI
include, rather than hardcoding this.

The navbar as copied to munstrap-navbar.tmpl is meant to be rendered
once and used in all pages. Therefore, rather than using R_PATH, which
should provide a relative path to the root depending on what is being
rendered, a fixed URI provided through a separate template variable
R_PATH_ABS is used for links pointing to home.